+++ index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Tetris</title>
   <link rel="stylesheet" href="style.css"></link>
-  <script defer src="script.js"></script>
+  <script defer type='module' src="script.js"></script>
 </head>
 <body>
   <header>
