+++ shapes/allshapes.js
@@ -1,0 +1,5 @@
+import { iPiece } from './ishape.js'
+import { oPiece } from './oshape.js'
+import { tPiece } from './tshape.js'
+
+export { iPiece, oPiece, tPiece }
