+++ shapes/oshape.js
@@ -1,0 +1,101 @@
+export class oPiece {
+  constructor() {
+    this.color = 'yellow'
+    this.active = true
+    this.center = [5, 1]
+
+    this.calculatePieces()
+  }
+
+  calculatePieces() {
+    this.bot = [ ...this.center ]
+    this.bot[1] = this.bot[1] + 1
+
+    this.botRight = [ ...this.center ]
+    this.botRight.forEach((coord, index) => {
+      this.botRight[index] = coord + 1
+    })
+
+    this.topRight = [ ...this.center ]
+    this.topRight[0] = this.topRight[0] + 1
+
+    this.tiles = [this.bot, this.botRight, this.center, this.topRight]
+  }
+
+  checkFall(grid) {
+    let check = 'false'
+    this.tiles.forEach(tile => {
+      let [x, y] = tile
+      y += 1
+      if(check === 'false') {
+        check = grid[y][x].getAttribute(['data-taken'])
+      }
+    })
+    if (check === 'true') {
+      this.tiles.forEach(tile => {
+        const [x, y] = tile
+        grid[y][x].dataset.taken = 'true'
+      })
+      this.active = false
+    } else {
+      this.fall()
+    }
+  }
+
+  fall() {
+    if (this.bot[1] < 10) {
+      this.center[1] += 1
+      this.calculatePieces()
+    }
+  }
+
+  checkRotate() {
+    this.rotate()
+  }
+
+  rotate() {
+    console.log("Nothing happens ye fool")
+  }
+
+  checkLeft(grid) {
+    let check = 'false'
+    this.tiles.forEach(tile => {
+      let [x, y] = tile
+      x -= 1
+      if(check === 'false') {
+        check = grid[y][x].getAttribute(['data-taken'])
+      }
+    })
+    if (check === 'false') {
+      this.left()
+    }
+  }
+
+  left() {
+    if (this.center[0] > 0) {
+      this.center[0] -= 1
+      this.calculatePieces()
+    }
+  }
+
+  checkRight(grid) {
+    let check = 'false'
+    this.tiles.forEach(tile => {
+      let [x, y] = tile
+      x += 1
+      if(check === 'false') {
+        check = grid[y][x].getAttribute(['data-taken'])
+      }
+    })
+    if (check === 'false') {
+      this.right()
+    }
+  }
+
+  right() {
+    if (this.topRight[0] < 10) {
+      this.center[0] += 1
+      this.calculatePieces()
+    }
+  }
+}
